Updated the demos so that they work with the new version of pyqtplot
